Refactor attributes
