+++ managed-ledger/src/main/java/org/apache/bookkeeper/mledger/ManagedCursor.java
@@ -34,7 +34,6 @@ import org.apache.bookkeeper.mledger.AsyncCallbacks.MarkDeleteCallback;
 import org.apache.bookkeeper.mledger.AsyncCallbacks.ReadEntriesCallback;
 import org.apache.bookkeeper.mledger.AsyncCallbacks.ReadEntryCallback;
 import org.apache.bookkeeper.mledger.AsyncCallbacks.SkipEntriesCallback;
-import org.apache.bookkeeper.mledger.impl.ManagedCursorAttributes;
 import org.apache.bookkeeper.mledger.impl.PositionImpl;
 
 /**
