+++ pulsar-opentelemetry/src/main/java/org/apache/pulsar/opentelemetry/OpenTelemetryAttributes.java
@@ -101,6 +101,7 @@ public interface OpenTelemetryAttributes {
      * The status of the Pulsar transaction.
      */
     AttributeKey<String> PULSAR_TRANSACTION_STATUS = AttributeKey.stringKey("pulsar.transaction.status");
+
     enum TransactionStatus {
         ACTIVE,
         COMMITTED,
@@ -129,6 +130,26 @@ public interface OpenTelemetryAttributes {
     }
 
     // Managed Ledger Attributes
+    /**
+     * The name of the managed ledger.
+     */
+    AttributeKey<String> ML_LEDGER_NAME = AttributeKey.stringKey("pulsar.managed_ledger.name");
+
+    /**
+     * The name of the managed cursor.
+     */
+    AttributeKey<String> ML_CURSOR_NAME = AttributeKey.stringKey("pulsar.managed_ledger.cursor.name");
+
+    /**
+     * The status of the managed cursor operation.
+     */
+    AttributeKey<String> ML_CURSOR_OPERATION_STATUS =
+            AttributeKey.stringKey("pulsar.managed_ledger.cursor.operation.status");
+    enum ManagedCursorOperationStatus {
+        SUCCESS,
+        FAILURE;
+        public final Attributes attributes = Attributes.of(ML_CURSOR_OPERATION_STATUS, name().toLowerCase());
+    }
 
     /**
      * The type of the pool arena.
